[ETCM-759] Move mining related classes to 'pow.miners'
